requirements.txt
fastapi
uvicorn
pandas
scikit-learn
sqlalchemy
pymysql
joblib
apscheduler
python-dotenv